add datepicker and changes

+++ src/app/cosmonauts/cosmonauts.component.js
@@ -21,8 +21,10 @@ var CosmonautsComponent = (function () {
     };
     CosmonautsComponent.prototype.deleteCosmonaut = function (cosmonaut) {
         var _this = this;
-        this.cosmonautService.deleteCosmonaut(cosmonaut.id)
-            .then(function () { return _this.cosmonauts = _this.cosmonauts.filter(function (cn) { return cn !== cosmonaut; }); });
+        if (confirm("Opravdu chcete smazat kosmonauta " + cosmonaut.name + ' ' + cosmonaut.surname + '?')) {
+            this.cosmonautService.deleteCosmonaut(cosmonaut.id)
+                .then(function () { return _this.cosmonauts = _this.cosmonauts.filter(function (cn) { return cn !== cosmonaut; }); });
+        }
     };
     return CosmonautsComponent;
 }());
